Split yaw into a dedicated module
